docs(fern): make Research publications its own sidebar entry
The papers were rendered as a second block on the External publications page,
which put them nowhere in the sidebar. They now have their own page nested
under External publications, so the tab reads:

  External publications
    Research publications
  Dynamo blogs
    ...

External publications becomes a section with `path:`, which keeps the partner
list reachable as the section's own overview -- the entry stays clickable and
/dynamo/dev/external-publications is unchanged.

The card CSS moves to PublicationsStyles.tsx, since two pages now need it and
it can no longer live inside either component. Naming it *Styles.tsx also puts
it back under check_style_components.py's own glob, so the hook's files pattern
and the script's EXTRA_TARGETS list drop the special case added in f07485b.
