+++ .pre-commit-config.yaml
@@ -110,10 +110,9 @@ repos:
       name: Check component CSS template literals
       entry: python3 docs/fern/scripts/check_style_components.py
       language: system
-      # EcosystemPublications.tsx holds its CSS inline rather than in a
-      # *Styles.tsx file, so it needs naming here to be covered. It is not the
-      # only component that does — see the script's docstring.
-      files: ^docs/fern/(components/(.*Styles|EcosystemPublications)\.tsx|scripts/check_style_components\.py)$
+      # Components that keep CSS outside a *Styles.tsx file need naming here to
+      # be covered — see the script's docstring. None do today.
+      files: ^docs/fern/(components/.*Styles\.tsx|scripts/check_style_components\.py)$
     - id: check-published-styles-selftest
       name: Self-test the published-styles regex helpers
       entry: python3 docs/fern/scripts/check_published_styles.py --test
