+++ docs/fern/pages/blog/research-publications.mdx
@@ -1,17 +1,17 @@
 ---
 # SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
-title: External publications
-description: Articles, deep dives, and deployment write-ups about NVIDIA Dynamo published by customers and partners.
+title: Research publications
+description: Papers that use, extend, or benchmark NVIDIA Dynamo, from the teams building on it and the wider systems community.
 hide-toc: true
 hide-page-actions: true
 ---
 
 import { BlogStyles } from "@/components/BlogStyles";
 import { PublicationsStyles } from "@/components/PublicationsStyles";
-import { EcosystemPublications } from "@/components/EcosystemPublications";
+import { ResearchPublications } from "@/components/ResearchPublications";
 
 <BlogStyles />
 <PublicationsStyles />
 
-<EcosystemPublications />
+<ResearchPublications />
